MOBILE-3320 behat: Fix change detection in ionic5
